pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui_image_saver"
version = "0.1.0"
description = "ComfyUI custom node for saving images without metadata"
authors = [
  {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.comfy]
PublisherId = "your_username"
DisplayName = "Image Saver (No Metadata)"
Icon = ""
Tags = ["image", "save", "metadata"]
Repository = "https://github.com/your_username/comfyui_image_saver"
includes = []

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.ruff]
line-length = 140
src = ["src", "tests"]
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",  # Pyflakes
    "W293",
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
